support optional chaining in unit test files
with a pointer from @raathigesh I was able to add support to babel which allows unit tests to contain statements that use the new optional chaining syntax.

This is necessary because Majestic processes the unit test file with a locally configured copy of babel. We had to add the experimental plugin for optional chaining to babel.

This fixes issue #164