added repeatable ID generation for jest elements
This is the first step to collapsable describe blocks in the UI. This change modifies the way that Ids are assigned to the jest blocks that are passed to the UI. Before this change, every time the file was refreshed, the Id assigned to a block would be brand new and different. With this change we keep track of the Ids that have been assigned and we re-use them when we recognize that the element is the same as before.

This will allow us in the UI to keep track of which describe blocks have been collapsed and when we revisit the page or rerun the tests, the describe blocks that were collapsed will remain collapsed.

This is just the server side of that change.