Added collapsing describe blocks
This checkin adds the ability to collapse the describe blocks for the test result. The collapsed state is stored per describe block in a static class and updated each time the user toggles a block. The collapse state storage is small, only storing data for the describe blocks that are actually toggled.
When a test is re-run, if there are failures in a describe block that is collapsed, it is automatically opened.
I also fixed the display of the status icon beside describe blocks. Before this change it was always green. Now it is red when one of its tests fail. I am pretty sure this was a bug before because as soon as I changed the status to failed for the describe block, it changed the colour to red, indicating that the testIndicator component already handled the case but wasn't being passed the correct values.

I believe this change fits in with the style of the existing component. I retained the functional component style by using React hooks. Unexpectantly, I had to use the useEffect tied to changing of the results variable. This was necessary to get the describe block to open after a test run if a test within it failed.

I also had to change the implementation of allChildrenPassing. it now recursively looks at the children where it didnt before.

Also, I made it so that clicking on the entire bar collapsed or expands the block rather than just the +/- div.