fixed issue where tests with the same name would share the same failure results
Since the UI was matching results with tests by the test title, it was picking the wrong test result for all but the the first test with the duplicated name.
This fix uses the previous change for collapsing describe blocks. The Ids that the backend assigned to the tests, are propogated through to the UI with the test results with this change. This allows the UI to use the id to match the test results with the tests. The end result is that a user can have the same named test in multiple locations in the test file and the UI will always match up the correct results with the tests.

The id field had to be added to the test item result entry and the queries the UI sends had to include the field.

Assigning Ids to the tests in the test result is fairly efficient because it can use the property of the test results that tests are grouped by under their describe block. So on each iteration of the loop, we first check to see if the existing idManaager is still valid and if it is, we just use that one rather than searching for the right idManager to use.