implement view snapshot button
This commit implements issue #186. I can't know if there is a snapshot for each
but I can add a button to the summary that opens the snapshot for the file
in the code editor. That's what this commit does. It adds the button to the
UI and a mutation to send to the server to open the snapshot in the editor.