changed default collapsed state to closed
One of the other majestic users commented on the issue asking if the state defaulted
to closed or open. Thinking it through, it should default to closed but the
implementation defaulted to open. So it has now been changed to default to closed
unless there is a failing test. Then it springs open to show the failed test.

The comments were mentioned in #186 but this change is really for #180