193 better indication test run status (#198)
* changed summary to only animate when updating

When in watch mode, the summary background used to animate all
the time. This meant that it was difficult to understand if the test was
currently running or if the results you were looking at were the result
of the latest test run.

With this change the background only animates when a test is actually
in progress. If you are in watch mode, the play button turns to a stop
button and the background is not animated, But when Jest is running,
the background is animating. When jest finishes, the background stops.

This addresses issue #193

* fixed merge error

cherry picking the commit brought in part of another change.
this undoes that other change and puts the removed code back.

Co-authored-by: Greg Veres <greg.veres@rogers.com>