Keeps the results of previous run when starting new run
These changes fetch the previous run's results when a new run starts.
This allows the user to keep looking at the previous results while watch
mode starts a new test run. This can save a lot of time because the
user can start looking at the next failure that needs to be fixed while
the tests are run on the latest change. Without these changes, the
previous run's results were discarded as soon as the new test run
started. That meant that the developer had to sit there waiting for
the results of the new run just to see the error condition on the next
test that was on screen before she pressed save.

I added typing to the 'results' map in the results.ts file to make the
results easier to work with.

(cherry picked from commit 5557405243c899231cda02094c4ad3f184476338)