dimmed the results information when test running
I wanted to provide an additional visual cue to the user that the results
they were looking at were out of date, but still alow them to view the
results while the other results were being generated.

Using the information from this SO question:
https://stackoverflow.com/questions/48502647/conditional-rendering-in-styled-components
I added a dim property to the Content panel that is triggered
on isUpdating.