Add collapse all and expand all to the tree control
I have added two icons to the tree control icon area. The first collapses
all the directories in the tree except the root directory and the other
expands all the directories in the tree.

This makes it much easier to find a test file to work on.
I also put the icons on the right side because that is where the
"show failing tests only" icon appears and it seemed that they were
similar types of operations. Then I realized that the collapse / expand
don't make sense when the show failing test only icon is toggled
so in that case I hide the icons.

And then because the icons vanish and appear, I rearranged their
position to what I thought made more sense.
Refresh: is always present and always in a fixed location on the right
The most important button is show only failed so it shows up on the
left as the first in the list when there are failed tests.
Then the collapse/expand
Then the coverage (which I think is the least likely to be used) and
therefore in the "hardest spot to hit"
#203